Rev_B_EPS Schematic Updates
EPS Board Schematic is starting to come to a conclusion. Swapped out to new load switches that use external mosfets and have an enable switch. R and C values finalized for almost everything. Need some reviewing and/or notes on how to proceed with Phases for each component plus if we might need an external DC/DC converter just for the clock oscillator. Good stuff