Calculations
Lots of behind the scenes stuff going on. Ran through calculations for the Load Switch, Switching Regulator and clock for the jetson using the information from each datasheet. Having to modify these parameters for each subsystem is time consuming as each one requires new calculations. Once this bit is done the schematic will be done, leading to layout. Aiming for thursday Layout.