Layout and Routing DONE
Layout and Routing is complete, next step is DRC checking and making sure that the tracks and fab settings coincide with the settings we want to manufacture with.  Worth noting that we should probably get a version of this board with just one subsystem to test first before buying the whole thing.