Added picture to create a block diagram
Placed pictures and wires with some text to give a block diagram.

Looked through the data sheets and found some info on the power for the ethernet switch. The power needs to go through a ferrite bead. Since we are getting power through a USB port I will use voltage regulators to get the 3.3V and the 1.2V needed for the board.

Tried to get the same version that we are using for the raspberry pi pico but the easyeda2kicad was not working. There is already one that was from the dev board that I have placed but I would like to keep the parts in the same family if possible.