Update to Schematic & PCB layout
Updated the schematic to have a non populated USB-C port to add in at a later date.  Removed some reference pages that I no longer using.

For the PCB I have been making edits to footprints to closely match the RP2350 design requirements. Making the placement of the parts exactly or close to that of the requirements and guide lines.

Have left the Wiznet alone and the Ethernet switch is not touched because it is mainly signal tracking.

Still need to make the changes for the signal integrity of the Ethernet switch from the data sheet.

Made sure the parts from my global library are include in the project specific one.