Update to PCB and minor changes to schematic
Completed the RP2350 layout. Put in the edge cut that should match the IMX flight board.

Had to rewire the Ethernet ports but finished one just need to copy and paste it to the others. Once that is done and connected to the Wiznet I can move to the switch.