Checked the schematic and made edits to PCB
Looked through Wiznet's GitHub to get a better understand how to connect it to RP2350.

Did the basic layout of the board with the RJ45 and the DF11.

The ethernet switch chip needs to be connected with its decoupling caps and then lastly the RJ45 need to be connected to their respective chips.