Updates to the Ethernet switch
Made a change to power to the ethernet switch. The original was thermally regulated so I moved to using one that was similar for the raspberry pi.
Started putting some of the resistors and LEDs for debugging on the chip.
Leaving the decoupling capacitors till later to make sure they are attached to the correct pins.
Needed to add some symbols and libraries back to the project.