Update to schematic and worked on PCB
Made a few changes with signals for the ethernet switch schematic. Made it easier for the traces for power.

Put all the decoupling capacitors onto the ethernet switch chip on the PCB. Was able to route the traces for the Wiznet. Might want to change where the signals are. Want to look into maybe making the 2 or 3 plane as a signal plane for just the ethernet to ensure direct connections.

Just need to connect the power of 1.2 volts that the ethernet switch needs and then will be able to route the 4 other ethernet switches. I am being sure to use the pair routing for the length.

Side note figured out how to add images as a layer, which is fun!