Update to the PCB
Did the signal paths for the Ethernet switch. Need to put in the pathing for the 1.2V analog and digital for the switch.

Need to finish some other signal connections for the DF11 and some of the test LEDs.

Once that is in place I will finish the fills and the board will be complete for a review.

*Disclaimer:* I am not too sure about the placement of the ethernet signals. I am thinking of taking them off the back plane and putting it on the dedicated ground plane.