Updated to PCB layout
Made changes to the layout. Made it so there are only 2 sides of ports for the RJ45 connectors.
Completed the BOM for the board. Leaving the JTAG as just pads for dev purposes.
Started the error checking for the the board. All but 2 errors and the rest warnings but mainly placement based.
Should be a cleaner version.