Board Layout complete
Its a bit messy but I think ive managed to get the loop areas small enough, and the board to represent what is on the schematic. Probably would be a good idea to get a look at it from someone more reputable than myself on how this board should be layed out.