Changes to the PCB
##Brief list of what I changed:
1. Changed all the via sizes to be 0.6/0.3 mm which lead to me having to reroute a some things

2. Changed the trace width to be a bit larger for some signals to see if that would be better instead of the smaller traces I had prior.

3. In rerouting I wanted to see if different configurations of the RJ45 connections would be better.

4. Modified some of the fill zones to eliminate traces.

##Overall summary:
Changed via and track sizes because JLC said that I had some errors on my board that need further changes/testing which increased the price. Some of these changes required me to reroute to where I wanted to try and use larger trace sizes to see if that would be better or not. The functionality should still be there. Most of my changes where to the ethernet switching part. If the boards are still ordered then this will be great for me to check and if they work as intended then I will revered the trace sizes and some of the routing. For two of the RJ45 I changed to have one of them be connected on the top layer and bottom layer while another was connected to the pins of the RJ45 and not directly next to the resistors.