Minor KiCad Changes
Combing through my KiCad Design and making sure that everything is up to standard.
Board seems to be a bit off and doesn't seem to to be working correctly.
There are a couple things to note.

For the Wiznet there are different configurations that seem to me that were unclear and require a closer look.

For the Ethernet Switch I need to make sure it is being power correctly and I might need to had in the other ethernet switch to ensure that it works as intended because the 5th port I left open but the switch does not seem to work.