Updated to the schematic
Kicad died on me so I needed to readd all my libraries to the project.

Added in the the Power board connector but want to go over it to make sure it will be implemented correctly. It is currently excluded.

Looked through the SALAD board and I had the same layout but I guess I did something wrong with selection of and or placement of signals traces. Added back in the single bus buffer, jumpers for changing the pmode and the filtering of the power and ground. It is currently excluded. Wiznet RJ45 connection is done.

Still going through the ethernet switch but found more info on the strap in mode. It is configured from the LED signals. The design guide has some specific details for the RJ45 and the Wiznet's RJ45 should be fine to use but I a Pulse JXD6-0001NL as a an easier stand in to better match the hardware design guideline. Still need to figure out the stand alone 4 port switch.