Schematic design ready for review
Made the last changes for the Wiznet by adding in Dip switches and made sure that the interrupt and reset signals are attached to the raspberry pi.

Finished the Etherent switch by wiring the RJ45 port 1 with 2 configurations, leaning more towards not using the Wiznet port but might do it any ways for cost. Added test points to port 5 but might change it to be RMII configuration because the 4 port strap in mode sets it to that config. Put solder jumpers in for the LED strap in mode. On the schematic I put in a note about how to configure to the proper mode we want but after a review will change to be a hardwire set. Made the reset switch match that of the EVAL board but further reading found I do not need the diode. Created the VDDIO and VDDAT which are just 2 different versions of 3.3V. Layout would be next but need to learn more about trace integrity and impedance control.