Started the PCB layout
Finished the schematic. Now working on the PCB. Might looking in to moving to a 6 layer board.
Need to do a bit of redesign because changing some of the parts had different footprints.

Started on the Wiznet and the its RJ45 port it has. Want to make the port close to the chip as a possible. Currently working on the 4 layer stack up but if I switch to a 6 layer it might help with impedance and change the overall signal layering.
Almost done with the Wiznet just need to do the clock and the TX/RX pairs.
Once that is done I will work on the Ethernet switch then lastly the breakout pins for the RP.

I should have selected either a 6 layer or 4 layer board, leaning more towards 6 layer because I can keep my signals separated by the 3 and 3 sections.