PCB layout
Added a new version of the board that went from 4 layers to 6.

Started the 6 layer board and moving forward with it.
Need to restructure the board and use the via in pads.
The wiznet is basically done just needs to be placed on the board.
The RP is almost done but the GPIO is broken out but still need to put in the pins.
Ethernet switch is mostly done. Need to figure out routing for the test pads and then place the RJ45 ports on the board. Once I have the ports and the chip in place there calculations have already been done for the routings just need to add the net class to the pins then it will be impedance protected.