PCB Update
Got all the parts on the board placed. Now just finishing the routing for each interconnection.
Trying to maintain a side to side/up to down but the board kinda follows it. For each plane  there is fill zones that will be for the chip that is on the top. Some of the traces are just being fed to where they need to go but the important EMI signals are being contained to their guidelines.
I will need to look at the lengths of the traces to make sure they are kept around the same length.

Will need to go over the board with some other eyes to make sure its ok.

I made sure the Wiznet traces are the same as SALAD.