PCB should be done
Finished the PCB just needs a review and then the part selection should be easy.
Added in stitching Vias.
Added in a Power controller for the two power sources of the USB or the external power board.
Removed the microSD card reader and added in XTSD which is just onboard NAND. It is around 1 gigabyte which should be enough space for any and all programs we are running.