PCB step update
Made some additions to the 3d files for the project.
Deleted some old files that were just not being used.
Would like to restructure the folders and remove the v1 and make a folder that is just KiCad dependancies to help version control.
*Unsure if that is a good idea or if I should just leave it as project specific libriaries.*
Will look over this before this is merged to main.