Working on the board layout
The differential pairs are being a pain. Signals B and D need to be crossed which is not ideal. Some of the signals do not seem to want to connect which is making this a fun time. Still need to put in the structures like the drill holes. I have not done the capacitors yet but I think I might do that next time then do the DP routing to make life a bit more straight forward.