Board layout WIP
Redid the placement of the board due to the structure which should be good now. Started to put stuff on the back of the board to make sure that there is enough space for the Ethernet signals. Have not put the pull configuration on the board yet but will do so after the the RJ45 connections are done. This still needs to be decided if they are placed near their port or the pin on the chip. Most of the tougher layout is done. Will still need to run through the design check list to make sure it satisfies the board.