Worked on IMX part of schematic
Did the big part placement for the IMX part of the board. Finished the boot config, debug, ethernet ports, ethernet power and ethernet trigger buffer.

Still need to do the major part of the ethernet IC.

There are a few comments on the schematic we got back from Phytec that I am unsure what they mean.