Updates from Microchip
Finally got a response for my schematic from Microchip support. Made a few changes that they recommended. Changed the crystal, updated a power pin, and added in the resistors for the LEDs. Sent it back with the updates and once it seems okay I will move everything over to layout and then send it to them there. I will need to make it in Altium to make life easier since they do not use KiCad.