Little Updates from Microchip
Got some feedback from microchip right before the Holiday weekend and implemented them on to the board and asked some clarifying questions on what steps to do moving forward with some design guidelines. Not much feedback on the layout but there was some confusion since it was ported from KiCad to Altium.