Made board a 6-layer
Microchip wanted the board to be a 6-layer stack up because they said a solid ground plane under the differential signals was needed. Waiting on them now. I might in the mean time look for an alterative for magnetics and use of the geckos but I do not think it will be needed/used.